// src/modules/Register/hooks/types.ts
export interface IReturnRegisterFormData {
  amount_of_milk_produced: string;
  farmerCity: string;
  farmerName: string;
  from: string;
  latitude: string;
  longitude: string;
  number_of_cows_head: string;
  to: string;
  type: string;
  has_supervision: string;
}  
// {"amount_of_milk_produced": "22", "farmerCity": "city", "farmerName": "farmer", "from": "fromname", "has_supervision": "true", "latitude": "055", "longitude": "0555", "number_of_cows_head": "10", "to": "tonome", "type": "name"}